Add files via upload
The IsValidEmail() function failed the IsValidEmail_Is_True_Given_Valid_Email test function on the 'spr0cket@example.com' email address. This is because the
regular expression string set in the ValidEmailAddressPattern only allowed the prefix of the email address to contain the digits 1-9. In the case of 'spr0cket@example.com' we see that there is a 0 character in the prefix, making the address invalid. With the assumption being that email addresses with 0 in the prefix should be considered valid since the unit test includes one, the bug was fixed by simply editing the regular expression string to allow for digits 0-9 in the prefix. I also made this change to the suffix to maintain consistancy.