Order Service Bug Patches
The OrderService tests were failing because the JSON data was being pulled incorrectly. In the case of the GetCustomerTotalPaidUsd() function, the test was returning $120 instead of the expected $60. This is because there was no conditional expression specifying that the order objects to be summed need to have a PaidStatus of "Paid". That is, only the customer.Id and the order.Bike.Owner.Id were being checked for equality. The bug occurred because the same customer had two orders, each for $60, but only one of the orders had a PaidStatus of "Paid" while the other had a PaidStatus of "Waived", meaning that it should not have been included in the summation. The bug was fixed by adding a conditional that requires that the order.PaidStatus must equal "Paid".

The GetAllOrdersForCustomer() test was failing because the customer.Id was being checked against the order.Bike.Id (that is, the ID of the bike itself) instead of the order.Bike.Owner.Id (that is, the ID of the customer who purchased the bike). The bug was fixed by simply changing the code to check the customer.Id against the correct property (order.Bike.Owner.Id).